Make the border transparent instead of removing it to stop the jiggle.
The reason the jiggle was happening is that we were changing the padding, but the grid cell is always a fixed size. So the text was getting squeezed when there was more padding and therefore slightly changing position. If we fix the aount of padding and clip the background color we can still have an outline with some separation from the content.